+++ fixtures/0001-source-summary/input/DecoratorFree.tsx
@@ -1,0 +1,7 @@
+import { Component } from "presolve";
+
+export default class Home extends Component {
+  render() {
+    return <main>Home</main>;
+  }
+}